pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drone-pufferlib"
version = "0.1.0"
description = "Vision-based RL navigation with Gymnasium, PyTorch, and PufferLib-compatible adapters."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "Cython>=3.0.0",
  "gym==0.23.0",
  "gymnasium>=0.29.1,<=0.29.1",
  "matplotlib>=3.8.0",
  "ninja>=1.11.1",
  "nvidia-ml-py>=12.560.30",
  "numpy>=1.26.0,<2.0",
  "opencv-python-headless==4.10.0.84",
  "pufferlib==3.0.0",
  "psutil>=5.9.0",
  "pyro-ppl>=1.9.1",
  "pyyaml>=6.0.1",
  "rich>=13.7.0",
  "rich-argparse>=1.5.2",
  "tensorboard>=2.18.0",
  "torch>=2.5.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[tool.setuptools.packages.find]
include = ["drone_pufferlib*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
